Preliminary pulseaudio libs and utils - WIP
A working conf file for pulseaudio clients is created, and the
pkg-message contains all the required post-installation steps.

Native pulseaudio is a dependecy for the linux-f20-pulseaudio-libs
and the latter is the dependecy for the pulseaudio-utils.

Currently you need to manually install any dependecy on linux-f20- port.
Will be fixed soon :)

This depends on a patch for lemul linux32, that will be pushed somewhere
in a bit :)

Signed-off-by: Vassilis Laganakos <vassilis@einval.com>