Preliminary ports for alsa-plugins-pulseaudio and alsa-utils.
alsa-utils to be used for testing with aplay the integration with
pulseaudio - flashplugin uses alsa.

It is *not* required to fiddle with /compat/linux/etc/asound.conf, since
it looks like that the config files come with the RPM package and they
are loaded from /compat/linux/usr/share/alsa/alsa.conf.d/ that point
alsa to use pulseaudio.

Signed-off-by: Vassilis Laganakos <vassilis@einval.com>